% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{corpus_description}
\alias{corpus_description}
\title{corpus_description}
\usage{
corpus_description(data, text)
}
\arguments{
\item{data}{data}

\item{text}{text variable}
}
\value{
out ...
}
\description{
corpus_description
}
