% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_lstm.R
\name{k_cnn_lstm}
\alias{k_cnn_lstm}
\title{k_cnn_lstm}
\usage{
k_cnn_lstm(in_dim = 10000, in_length = 100, embed_dim = 128,
  sp_drop = 0.2, filter_sizes = c(1, 2, 3, 5), num_filters = 100,
  kernel_size = 5, pool_size = 4, lstm_dim = 64, lstm_drop = 0.3,
  layer_drop = 0.3, out_dim = 1, out_fun = "sigmoid", ...)
}
\arguments{
\item{in_dim}{Number of total vocabluary/words used}

\item{in_length}{Length of the input sequences}

\item{embed_dim}{Number of word vectors}

\item{sp_drop}{Spatial Dropout after Embedding}

\item{filter_sizes}{Filter sizes (windows)}

\item{num_filters}{Number of filters per layer}

\item{kernel_size}{Size of feature map}

\item{pool_size}{Size of feature map}

\item{lstm_dim}{Number of neurons in the LSTM layer}

\item{lstm_drop}{Dropout ratio in the LSTM layer}

\item{layer_drop}{Dropout ratio between layers}

\item{out_dim}{Number of neurons of the output layer}

\item{out_fun}{Output activation function}

\item{...}{Exit arguments}
}
\value{
model
}
\description{
get Keras stacked CNN LSTM model
}
\examples{

}
