% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glove.R
\name{k_glove}
\alias{k_glove}
\title{k_glove}
\usage{
k_glove(in_dim = 10000, in_length = 100, embed_dim = 128, out_dim = 1,
  out_fun = "sigmoid", ...)
}
\arguments{
\item{in_dim}{Number of total vocabluary/words used}

\item{in_length}{Length of the input sequences}

\item{embed_dim}{Number of word vectors}

\item{out_dim}{Number of neurons of the output layer}

\item{out_fun}{Output activation function}

\item{...}{Exit arguments}
}
\value{
model
}
\description{
get Keras GloVe model
}
