% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gru_cnn.R
\name{k_gru_cnn}
\alias{k_gru_cnn}
\title{k_gru_cnn}
\usage{
k_gru_cnn(in_dim = 10000, in_length = 100, embed_dim = 128,
  sp_drop = 0.2, gru_dim = 64, gru_drop = 0.2, out_dim = 1,
  out_fun = "sigmoid", ...)
}
\arguments{
\item{in_dim}{Number of total vocabluary/words used}

\item{in_length}{Length of the input sequences}

\item{embed_dim}{Number of word vectors}

\item{sp_drop}{Spatial Dropout after Embedding}

\item{gru_dim}{Number of GRU neurons}

\item{out_dim}{Number of neurons of the output layer}

\item{out_fun}{Output activation function}

\item{...}{Exit arguments}
}
\value{
model
}
\description{
get Keras stacked CNN GRU model
}
\examples{
Taken from * [Code](https://www.kaggle.com/mosnoiion/two-rnn-cnn-columns-networks-with-keras)
[Paper](http://people.idsia.ch/~ciresan/data/cvpr2012.pdf)

}
