% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm.R
\name{k_lstm}
\alias{k_lstm}
\title{k_lstm}
\usage{
k_lstm(in_dim = 10000, in_length = 100, embed_dim = 128, lstm_dim = 64,
  lstm_drop = 0.3, layer_drop = 0.3, out_dim = 1, out_fun = "sigmoid",
  ...)
}
\arguments{
\item{in_dim}{Number of total vocabluary/words used}

\item{in_length}{Length of the input sequences}

\item{embed_dim}{Number of word vectors}

\item{lstm_dim}{Number of neurons in the LSTM layer}

\item{lstm_drop}{Dropout ratio in the LSTM layer}

\item{layer_drop}{Dropout ratio between layers}

\item{out_dim}{Number of neurons of the output layer}

\item{out_fun}{Output activation function}

\item{...}{Exit arguments}
}
\value{
model
}
\description{
get Keras LSTM model
}
