% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{k_onehot}
\alias{k_onehot}
\title{k_onehot}
\usage{
k_onehot(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
transforms a k-level factor vector into a n*k matrix (one-hote encoding)\code{x}
}
\description{
k_onehot
}
