% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{params_helper}
\alias{params_helper}
\title{params_helper}
\usage{
params_helper(input, expend_params = F)
}
\arguments{
\item{input}{list}

\item{expend_params}{Wheter to grid expand the model params}
}
\value{
input modified
}
\description{
params_helper
}
