% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_text.R
\name{prepare_text}
\alias{prepare_text}
\title{prepare_text}
\usage{
prepare_text(data, in_dim = 10000, in_length = 100, char_level = F,
  return_tokenizer = F, ...)
}
\arguments{
\item{data}{list constisting of train and test set}

\item{in_dim}{Max top words (max_features)}

\item{in_length}{Max sequence length}

\item{char_level}{Word or character level?}

\item{...}{Exit arguments}
}
\value{
model
}
\description{
prepare_text
}
